+++ packages/service-library/tests/fastapi/long_running_tasks/test_long_running_tasks_context_manager.py
@@ -13,20 +13,18 @@ from pydantic import AnyHttpUrl, PositiveFloat, TypeAdapter
 from servicelib.fastapi.long_running_tasks._context_manager import _ProgressManager
 from servicelib.fastapi.long_running_tasks.client import Client, periodic_task_result
 from servicelib.fastapi.long_running_tasks.client import setup as setup_client
-from servicelib.fastapi.long_running_tasks.server import (
-    TaskId,
-    TaskProgress,
-    TasksManager,
-    get_tasks_manager,
-)
+from servicelib.fastapi.long_running_tasks.server import get_tasks_manager
 from servicelib.fastapi.long_running_tasks.server import setup as setup_server
-from servicelib.fastapi.long_running_tasks.server import (
-    start_task,
-)
 from servicelib.long_running_tasks.errors import (
     TaskClientTimeoutError,
 )
-from servicelib.long_running_tasks.models import ProgressMessage, ProgressPercent
+from servicelib.long_running_tasks.models import (
+    ProgressMessage,
+    ProgressPercent,
+    TaskId,
+    TaskProgress,
+)
+from servicelib.long_running_tasks.task import TasksManager, start_task
 
 TASK_SLEEP_INTERVAL: Final[PositiveFloat] = 0.1
