+++ services/director-v2/src/simcore_service_director_v2/api/routes/dynamic_scheduler.py
@@ -4,18 +4,15 @@ from typing import Annotated, Final
 from fastapi import APIRouter, Depends, HTTPException, status
 from models_library.projects_nodes_io import NodeID
 from pydantic import BaseModel, PositiveInt
-from servicelib.fastapi.long_running_tasks.server import (
-    TaskAlreadyRunningError,
-    TaskId,
-    TaskProgress,
-    TasksManager,
-    get_tasks_manager,
-    start_task,
-)
+from servicelib.fastapi.long_running_tasks.server import get_tasks_manager
+from servicelib.long_running_tasks.errors import TaskAlreadyRunningError
 from servicelib.long_running_tasks.models import (
     ProgressMessage,
     ProgressPercent,
+    TaskId,
+    TaskProgress,
 )
+from servicelib.long_running_tasks.task import TasksManager, start_task
 from tenacity import retry
 from tenacity.before_sleep import before_sleep_log
 from tenacity.retry import retry_if_result
