+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/api/rest/containers_long_running_tasks.py
@@ -2,14 +2,11 @@ from textwrap import dedent
 from typing import Annotated, cast
 
 from fastapi import APIRouter, Depends, FastAPI, Request, status
-from servicelib.fastapi.long_running_tasks.server import (
-    TaskAlreadyRunningError,
-    TaskId,
-    TasksManager,
-    get_tasks_manager,
-    start_task,
-)
+from servicelib.fastapi.long_running_tasks.server import get_tasks_manager
 from servicelib.fastapi.requests_decorators import cancel_on_disconnect
+from servicelib.long_running_tasks.errors import TaskAlreadyRunningError
+from servicelib.long_running_tasks.models import TaskId
+from servicelib.long_running_tasks.task import TasksManager, start_task
 
 from ...core.settings import ApplicationSettings
 from ...models.schemas.application_health import ApplicationHealth
